tasklist

+++ app/Http/Controllers/TasksController.php
@@ -78,6 +78,7 @@ class TasksController extends Controller
      */
     public function show($id)
     {
+        $this->authorize('view', $task); // 追加
         $task = Task::find($id);
 
         return view('tasks.show', [
@@ -115,6 +116,7 @@ class TasksController extends Controller
             'status' => 'required|max:10',
             'content' => 'required|max:191',
         ]);
+        $this->authorize('view', $task); // 追加
         $task = Task::find($id);
         $user = \Auth::id();
         //$task->user_id = $request->user_id;
@@ -134,7 +136,9 @@ class TasksController extends Controller
      */
     public function destroy($id)
     {
+        $this->authorize('view', $task); // 追加
         $task = Task::find($id);
+        
         if (\Auth::id() === $task->user_id) {
             $task->delete();
         }
